feat(breadcrumbs): split the example app, placing the native error types on a separate screen
