+++ example/lib/main.dart
@@ -1,11 +1,9 @@
 import 'dart:async';
-import 'dart:ffi';
-import 'dart:io';
 
 import 'package:bugsnag_flutter/bugsnag.dart';
 import 'package:bugsnag_flutter/widgets.dart';
+import 'package:examples/native_crashes.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 import 'bad_widget.dart';
 
@@ -29,7 +27,21 @@ void main() async => bugsnag.start(
 class ExampleApp extends StatelessWidget {
   const ExampleApp({Key? key}) : super(key: key);
 
-  static const _methodChannel = MethodChannel('com.bugsnag.example/channel');
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      navigatorObservers: [BugsnagNavigatorObserver()],
+      initialRoute: '/',
+      routes: {
+        '/': (context) => const ExampleHomeScreen(),
+        '/native-crashes': (context) => const NativeCrashesScreen(),
+      },
+    );
+  }
+}
+
+class ExampleHomeScreen extends StatelessWidget {
+  const ExampleHomeScreen({Key? key}) : super(key: key);
 
   // Unhandled exceptions will automatically be detected and reported.
   // They are displayed with an 'Error' severity on the dashboard.
@@ -53,24 +65,6 @@ class ExampleApp extends StatelessWidget {
     }
   }
 
-  // Crashes in native code will be reported when Bugsnag is next started.
-  void _nativeCrash() {
-    final dynamicLibrary = Platform.isAndroid
-        ? DynamicLibrary.open('libc.so')
-        : DynamicLibrary.process();
-    // Intentionally incorrect function definition + call that causes a crash
-    final int Function(int arg) strlen = dynamicLibrary
-        .lookup<NativeFunction<Int32 Function(Int32)>>('strlen')
-        .asFunction();
-    strlen(0);
-  }
-
-  Future<void> _anr() => _methodChannel.invokeMethod('anr');
-
-  Future<void> _fatalAppHang() => _methodChannel.invokeMethod('fatalAppHang');
-
-  Future<void> _oom() => _methodChannel.invokeMethod('oom');
-
   // Use leaveBreadcrumb() to log potentially useful events in order to
   // understand what happened in your app before each error.
   void _leaveBreadcrumb() async =>
@@ -80,78 +74,59 @@ class ExampleApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Bugsnag example app'),
-        ),
-        body: Center(
-          child: Column(
-            children: [
-              Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Text(
-                  'Unhandled Errors',
-                  style: Theme.of(context).textTheme.headline6,
-                ),
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Bugsnag example app'),
+      ),
+      body: Center(
+        child: Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Text(
+                'Unhandled Errors',
+                style: Theme.of(context).textTheme.headline6,
               ),
-              ElevatedButton(
-                onPressed: _unhandledFlutterError,
-                child: const Text('Unhandled Error'),
+            ),
+            ElevatedButton(
+              onPressed: _unhandledFlutterError,
+              child: const Text('Unhandled Error'),
+            ),
+            ElevatedButton(
+              onPressed: _asyncUnhandledFlutterError,
+              child: const Text('Async Unhandled Error'),
+            ),
+            ElevatedButton(
+              onPressed: _handledException,
+              child: const Text('Notify Handled Error'),
+            ),
+            ErrorBoundary(
+              child: const BadWidget(),
+              fallback: (context) => const Text(
+                'A build() error has occurred and been reported',
               ),
-              ElevatedButton(
-                onPressed: _asyncUnhandledFlutterError,
-                child: const Text('Async Unhandled Error'),
+            ),
+            ElevatedButton(
+              child: const Text('Native Errors'),
+              onPressed: () {
+                Navigator.pushNamed(context, '/native-crashes');
+              },
+              style: ElevatedButton.styleFrom(
+                primary: Colors.redAccent.shade200,
               ),
-              ElevatedButton(
-                onPressed: _handledException,
-                child: const Text('Notify Handled Error'),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Text(
+                'Breadcrumbs',
+                style: Theme.of(context).textTheme.headline6,
               ),
-              ErrorBoundary(
-                child: const BadWidget(),
-                fallback: (context) => const Text(
-                  'A build() error has occurred and been reported',
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Text(
-                  'Native Errors',
-                  style: Theme.of(context).textTheme.headline6,
-                ),
-              ),
-              ElevatedButton(
-                onPressed: _nativeCrash,
-                child: const Text('Native crash'),
-              ),
-              if (Platform.isAndroid)
-                ElevatedButton(
-                  onPressed: _anr,
-                  child: const Text('Application Not Responding (ANR)'),
-                ),
-              if (Platform.isIOS)
-                ElevatedButton(
-                  onPressed: _fatalAppHang,
-                  child: const Text('Fatal App Hang'),
-                ),
-              if (Platform.isIOS)
-                ElevatedButton(
-                  onPressed: _oom,
-                  child: const Text('Out Of Memory'),
-                ),
-              Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Text(
-                  'Breadcrumbs',
-                  style: Theme.of(context).textTheme.headline6,
-                ),
-              ),
-              ElevatedButton(
-                onPressed: _leaveBreadcrumb,
-                child: const Text('Leave a breadcrumb'),
-              ),
-            ],
-          ),
+            ),
+            ElevatedButton(
+              onPressed: _leaveBreadcrumb,
+              child: const Text('Leave a breadcrumb'),
+            ),
+          ],
         ),
       ),
     );
